cloud_run/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
google-cloud-firestore==2.14.0
google-cloud-tasks==2.15.0
google-genai==0.4.0
playwright==1.41.0
